Fix command line args for test player
Code is expecting the argument to be passed as 'key:value' at line 245 of Assets\OpenCog Assets\Scripts\OpenCog\Utility\OCConfig.cs. Because the arguments had an unexpected format, the following lines are in output_log.txt:
Ignoring unknown command-line argument '-test'.
Ignoring unknown command-line argument 'internal_XGA'.
Ignoring unknown command-line argument '-quit'.

Because of this, the player is not terminating. This caused the BuildBot to fail as reflected in the log file:
http://buildbot.opencog.org:8010/builders/unity3d_game/builds/17/steps/test/logs/stdio
command timed out: 1200 seconds without output, attempting to kill

I tried it with the new args and it terminated successfully.